Categoryを別ファイルに記述する+Dateformatは共通なのでstatic定数としてDateUtilsの中に定義した
